requirements.txt
fastapi>=0.110.0
starlette>=0.36.3
uvicorn>=0.21.1
python-dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
pydantic[email]>=1.10.7
asyncpg>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
httpx>=0.24.1
aiohttp>=3.9.0
python-multipart>=0.0.6
pytz>=2023.3
python-dateutil>=2.8.2
python-slugify>=8.0.1
typing-extensions>=4.5.0
urllib3>=1.26.15,<2.1.0
certifi>=2023.5.7
brotli
